chore: add guard against deleting experiment without notice

+++ src/prescan/experiment_guard.cpp
@@ -9,7 +9,13 @@ namespace symaware {
 ExperimentGuard::ExperimentGuard(prescan::api::experiment::Experiment& experiment) : experiment_(experiment) {
   const std::string filename{std::filesystem::current_path().filename().string() + ".pb"};
   if (std::filesystem::exists(filename)) {
-    fmt::print("WARNING: Removing existing experiment file: {}\n", filename);
+    fmt::println("WARNING: Existing experiment file: {}. Remove and continue? [y/N]", filename);
+    std::string response;
+    std::cin >> response;
+    if (response != "y") {
+      fmt::println("Exiting...");
+      exit(1);
+    }
     std::filesystem::remove(filename);
   }
   experiment.saveToFile(filename);
